amp-carousel: Remove BaseSlides
