+++ extensions/amp-carousel/0.1/slidescroll.js
@@ -16,7 +16,7 @@
 
 import {ActionTrust} from '#core/constants/action-constants';
 import {Animation} from '../../../src/animation';
-import {BaseSlides} from './base-slides';
+import {dev, user, userAssert} from '../../../src/log';
 import {Keys} from '#core/constants/key-codes';
 import {Services} from '#service';
 import {bezierCurve} from '#core/data-structures/curve';
@@ -25,7 +25,6 @@ import {
   realChildElements,
 } from '#core/dom/query';
 import {createCustomEvent, listen} from '../../../src/event-helper';
-import {dev, user} from '../../../src/log';
 import {dict} from '#core/types/object';
 import {dispatchCustomEvent} from '#core/dom';
 import {getStyle, setStyle} from '#core/dom/style';
@@ -42,6 +41,7 @@ import {
   unobserveWithSharedInOb,
 } from '#core/dom/layout/viewport-observer';
 import {triggerAnalyticsEvent} from '../../../src/analytics';
+import {BaseCarousel} from './base-carousel';
 
 /** @const {string} */
 const SHOWN_CSS_CLASS = 'i-amphtml-slide-item-show';
@@ -63,7 +63,7 @@ const CUSTOM_SNAP_TIMEOUT = 100;
 
 const TAG = 'AMP-CAROUSEL';
 
-export class AmpSlideScroll extends BaseSlides {
+export class AmpSlideScroll extends BaseCarousel {
   /** @param {!AmpElement} element */
   constructor(element) {
     super(element);
@@ -94,6 +94,33 @@ export class AmpSlideScroll extends BaseSlides {
 
     /** @private {boolean} */
     this.isTouching_ = false;
+
+    /** @private {?number} */
+    this.autoplayTimeoutId_ = null;
+
+    /** @private {boolean} */
+    this.hasLoop_ = false;
+
+    /** @private {boolean} */
+    this.loopAdded_ = false;
+
+    /** @private {boolean} */
+    this.hasAutoplay_ = false;
+
+    /** @private {number} */
+    this.autoplayDelay_ = 5000;
+
+    /** @protected {?number} */
+    this.autoplayLoops_ = null;
+
+    /** @protected {number} */
+    this.loopsMade_ = 0;
+
+    /** @protected {boolean} */
+    this.shouldLoop = false;
+
+    /** @private {boolean} */
+    this.shouldAutoplay_ = false;
 
     /**
      * 0 - not in an elastic state.
@@ -162,6 +189,42 @@ export class AmpSlideScroll extends BaseSlides {
   }
 
   /** @override */
+  buildCarousel() {
+    this.hasLoop_ = this.element.hasAttribute('loop');
+
+    this.hasAutoplay_ = this.element.hasAttribute('autoplay');
+    const autoplayVal = this.element.getAttribute('autoplay');
+    if (autoplayVal) {
+      this.autoplayLoops_ = parseInt(autoplayVal, 10);
+      userAssert(isFiniteNumber(this.autoplayLoops_));
+    }
+    this.buildSlides();
+
+    this.shouldLoop = this.hasLoop_ && this.isLoopingEligible();
+
+    this.shouldAutoplay_ = this.hasAutoplay_ && this.isLoopingEligible();
+
+    if (this.shouldAutoplay_ && this.autoplayLoops_ != 0) {
+      this.setupAutoplay_();
+    }
+
+    this.registerAction(
+      'toggleAutoplay',
+      (invocation) => {
+        const {args} = invocation;
+        if (args && args['toggleOn'] !== undefined) {
+          this.toggleAutoplay_(args['toggleOn']);
+        } else {
+          this.toggleAutoplay_(!this.hasAutoplay_);
+        }
+      },
+      ActionTrust.LOW
+    );
+  }
+
+  /**
+   * Builds slides
+   */
   buildSlides() {
     this.vsync_ = this.getVsync();
     this.action_ = Services.actionServiceForDoc(this.element);
@@ -269,7 +332,10 @@ export class AmpSlideScroll extends BaseSlides {
     unobserveContentSize(this.element, this.onResized_);
   }
 
-  /** @override */
+  /**
+   * Checks if a carousel is eligible to loop, regardless of the loop attribute.
+   * @return {boolean}
+   */
   isLoopingEligible() {
     return this.noOfSlides_ > 1;
   }
@@ -287,8 +353,29 @@ export class AmpSlideScroll extends BaseSlides {
    * @private
    */
   touchMoveHandler_() {
-    this.clearAutoplay();
+    this.clearAutoplay_();
     this.isTouching_ = true;
+  }
+
+  /** @override */
+  viewportCallbackTemp(inViewport) {
+    super.viewportCallbackTemp(inViewport);
+    if (inViewport) {
+      this.autoplay_();
+    } else {
+      this.clearAutoplay_();
+    }
+  }
+
+  /** @override */
+  goCallback(dir, animate, opt_autoplay) {
+    const trust = opt_autoplay ? ActionTrust.LOW : ActionTrust.HIGH;
+    this.moveSlide(dir, animate, trust);
+    if (opt_autoplay) {
+      this.autoplay_();
+    } else {
+      this.clearAutoplay_();
+    }
   }
 
   /**
@@ -404,7 +491,12 @@ export class AmpSlideScroll extends BaseSlides {
     return this.shouldLoop || this.slideIndex_ < this.slides_.length - 1;
   }
 
-  /** @override */
+  /**
+   * Proceeds to the next slide in the desired direction.
+   * @param {number} dir -1 or 1
+   * @param {boolean} animate
+   * @param {!ActionTrust} trust
+   */
   moveSlide(dir, animate, trust) {
     if (this.slideIndex_ !== null) {
       const hasNext = this.hasNext();
@@ -761,7 +853,7 @@ export class AmpSlideScroll extends BaseSlides {
     if (this.autoplayLoops_ && this.slideIndex_ === this.noOfSlides_ - 1) {
       this.loopsMade_++;
       if (this.loopsMade_ == this.autoplayLoops_) {
-        this.removeAutoplay();
+        this.removeAutoplay_();
       }
     }
     this.hideRestOfTheSlides_(showIndexArr);
@@ -926,5 +1018,105 @@ export class AmpSlideScroll extends BaseSlides {
    */
   analyticsEvent_(eventType, vars) {
     triggerAnalyticsEvent(this.element, eventType, vars);
+  }
+
+  /**
+   * Sets up the `autoplay` configuration.
+   * @private
+   */
+  setupAutoplay_() {
+    const delayValue = Number(this.element.getAttribute('delay'));
+    // If it isn't a number and is not greater than 0 then don't assign
+    // and use the default.
+    if (delayValue > 0) {
+      // Guard against autoplayValue that is lower than 1s to prevent
+      // people from crashing the runtime with providing very low delays.
+      this.autoplayDelay_ = Math.max(1000, delayValue);
+    }
+
+    // By default `autoplay` should also mean that the current carousel slide
+    // is looping. (to be able to advance past the last item)
+    if (!this.hasLoop_) {
+      this.element.setAttribute('loop', '');
+      this.loopAdded_ = true;
+      this.hasLoop_ = true;
+      this.shouldLoop = true;
+    }
+  }
+
+  /**
+   * Starts the autoplay delay if allowed.
+   * @private
+   */
+  autoplay_() {
+    if (!this.shouldAutoplay_ || this.autoplayLoops_ == 0) {
+      return;
+    }
+    this.clearAutoplay_();
+    this.autoplayTimeoutId_ = /** @type {number} */ (
+      Services.timerFor(this.win).delay(
+        this.go.bind(
+          this,
+          /* dir */ 1,
+          /* animate */ true,
+          /* autoplay */ true
+        ),
+        this.autoplayDelay_
+      )
+    );
+  }
+
+  /**
+   * Called by toggleAutoplay action to toggle the autoplay feature.
+   * @param {boolean} toggleOn
+   * @private
+   */
+  toggleAutoplay_(toggleOn) {
+    if (toggleOn == this.shouldAutoplay_) {
+      return;
+    }
+
+    const prevAutoplayStatus = this.shouldAutoplay_;
+
+    this.hasAutoplay_ = toggleOn;
+    this.shouldAutoplay_ = this.hasAutoplay_ && this.isLoopingEligible();
+
+    if (!prevAutoplayStatus && this.shouldAutoplay_) {
+      this.setupAutoplay_();
+    }
+
+    if (this.shouldAutoplay_) {
+      this.autoplay_();
+    } else {
+      this.clearAutoplay_();
+    }
+  }
+
+  /**
+   * Clear the autoplay timer.
+   * @private
+   */
+  clearAutoplay_() {
+    if (this.autoplayTimeoutId_ !== null) {
+      Services.timerFor(this.win).cancel(this.autoplayTimeoutId_);
+      this.autoplayTimeoutId_ = null;
+    }
+  }
+
+  /**
+   * Remove autoplay.
+   * @private
+   */
+  removeAutoplay_() {
+    this.clearAutoplay_();
+    if (this.loopAdded_) {
+      // Only remove if specified due to the `autoplay` attribute
+      this.element.removeAttribute('loop');
+      this.loopAdded_ = false;
+      this.hasLoop_ = false;
+      this.shouldLoop = false;
+    }
+    this.hasAutoplay_ = false;
+    this.shouldAutoplay_ = this.hasAutoplay_ && this.isLoopingEligible();
   }
 }
