+++ extensions/amp-carousel/0.1/test/test-slidescroll-slides.js
@@ -14,25 +14,9 @@
  * limitations under the License.
  */
 
-/**
- * Copyright 2016 The AMP HTML Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS-IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+import {AmpSlideScroll} from '../slidescroll';
 
-import {BaseSlides} from '../base-slides';
-
-describes.fakeWin('BaseSlides', {amp: true}, (env) => {
+describes.fakeWin('AmpSlideScroll', {amp: true}, (env) => {
   let win, doc;
   let buildSlidesSpy;
   let viewportCallbackSpy;
@@ -56,18 +40,27 @@ describes.fakeWin('BaseSlides', {amp: true}, (env) => {
     hasPrevSpy = env.sandbox.spy();
     hasNextSpy = env.sandbox.spy();
     goCallbackSpy = env.sandbox.spy();
-    setupAutoplaySpy = env.sandbox.spy(BaseSlides.prototype, 'setupAutoplay_');
-    buildButtonsSpy = env.sandbox.spy(BaseSlides.prototype, 'buildButtons');
-    setupGesturesSpy = env.sandbox.spy(BaseSlides.prototype, 'setupGestures');
+    setupAutoplaySpy = env.sandbox.spy(
+      AmpSlideScroll.prototype,
+      'setupAutoplay_'
+    );
+    buildButtonsSpy = env.sandbox.spy(AmpSlideScroll.prototype, 'buildButtons');
+    setupGesturesSpy = env.sandbox.spy(
+      AmpSlideScroll.prototype,
+      'setupGestures'
+    );
     setControlsStateSpy = env.sandbox.spy(
-      BaseSlides.prototype,
+      AmpSlideScroll.prototype,
       'setControlsState'
     );
-    hintControlsSpy = env.sandbox.spy(BaseSlides.prototype, 'hintControls');
-    autoplaySpy = env.sandbox.spy(BaseSlides.prototype, 'autoplay_');
-    clearAutoplaySpy = env.sandbox.spy(BaseSlides.prototype, 'clearAutoplay');
+    hintControlsSpy = env.sandbox.spy(AmpSlideScroll.prototype, 'hintControls');
+    autoplaySpy = env.sandbox.spy(AmpSlideScroll.prototype, 'autoplay_');
+    clearAutoplaySpy = env.sandbox.spy(
+      AmpSlideScroll.prototype,
+      'clearAutoplay_'
+    );
     viewportCallbackSpy = env.sandbox.spy(
-      BaseSlides.prototype,
+      AmpSlideScroll.prototype,
       'viewportCallbackTemp'
     );
   });
@@ -91,7 +84,7 @@ describes.fakeWin('BaseSlides', {amp: true}, (env) => {
     return element;
   }
 
-  class TestCarousel extends BaseSlides {
+  class TestCarousel extends AmpSlideScroll {
     /** @override */
     buildSlides() {
       buildSlidesSpy();
@@ -298,7 +291,7 @@ describes.fakeWin('BaseSlides', {amp: true}, (env) => {
     expect(clearAutoplaySpy).to.have.been.called;
     expect(carousel.autoplayTimeoutId_).to.not.be.null;
 
-    carousel.clearAutoplay();
+    carousel.clearAutoplay_();
     expect(carousel.autoplayTimeoutId_).to.be.null;
   });
